# openapi.yaml
openapi: 3.0.3
info:
  title: Vision Financial API
  version: "1.0"
  description: >
    Self-hosted financial transaction manager.
    All responses are wrapped in the ADR-026 envelope:
    `{ ok: true, data: T }` on success, `{ ok: false, error: {...} }` on failure.
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Shared envelope & list wrappers ─────────────────────────────────────────
components:
...
  schemas:
...
    Envelope:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        data:
          description: Present on success
        error:
          $ref: "#/components/schemas/ErrorPayload"
    ErrorPayload:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
...
    RecipientBankAccount:
      type: object
      required: [id, recipient_id, iban, is_primary]
      properties:
        id:
          type: integer
        recipient_id:
          type: integer
        iban:
          type: string
        bic:
          type: string
        bank_name:
          type: string
        is_primary:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ─── Market ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/recipients/{id}/bank-accounts:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    get:
      summary: List bank accounts for recipient
      operationId: getRecipientBankAccounts
      tags: [Recipients]
      responses:
        "200":
          description: Bank account list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RecipientBankAccount"
    post:
      summary: Add bank account to recipient
      operationId: createRecipientBankAccount
      tags: [Recipients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [iban]
              properties:
                iban:
                  type: string
                bic:
                  type: string
                bank_name:
                  type: string
                is_primary:
                  type: boolean
      responses:
        "201":
          description: Created bank account
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecipientBankAccount"
  /api/recipients/{id}/bank-accounts/{accountId}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
      - in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    patch:
      summary: Update bank account
      operationId: updateRecipientBankAccount
      tags: [Recipients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                iban:
                  type: string
                bic:
                  type: string
                bank_name:
                  type: string
      responses:
        "200":
          description: Updated bank account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
    delete:
      summary: Deactivate bank account (soft delete)
      operationId: deleteRecipientBankAccount
      tags: [Recipients]
      responses:
        "200":
          description: The deactivated bank account (soft delete — the row survives)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
  /api/recipients/{id}/bank-accounts/{accountId}/set-primary:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
      - in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
    post:
      summary: Set bank account as primary
      operationId: setRecipientBankAccountPrimary
      tags: [Recipients]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"

  # ─── Planned Transactions ────────────────────────────────────────────────────
...